Catch errors in Texture data
Either texture issues or map corruption often causes failures in getting textures.
Previously this would result in entire chunks being unrendered and it was difficult to figure out why.

By catching the failures here, it is easy to see exactly which block is causing issues.

It also allows the chunk to continue to render with one missing block, instead of the entire chunk being blank.