Properly Render invisible blocks
I found two block in my world: invisibleBedrock and movingBlock that were failing to render and causing "black holes".

invisibleBedrock is something one can get via a command stick into the world. It is essentially what it says.

I'm not sure what movingBlocks are. Possibly the result of a bug in old versions.

I have marked both with a structure_void block that gives a red block around the spot, and also set them to transparent so you an see the terrain under them.